css-text/hyphens/soft-hyphens-000.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8">
<title>soft hyphens, hidden by default</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://dev.w3.org/csswg/css-text-3/#word-break-property'>
<meta name="assert" content="When the word is not broken the UA will not display a hyphen-like glyph inside a word containing soft hyphens.">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style type='text/css'>
@font-face {
	font-family: 'csstest_asciiregular';
	src: url('support/csstest-ascii.eot');
	src: url('support/csstest-ascii.eot?#iefix') format('embedded-opentype'),
		 url('support/csstest-ascii.woff2') format('woff2'),
		 url('support/csstest-ascii.woff') format('woff'),
		 url('support/csstest-ascii.ttf') format('truetype');
	font-weight: normal;
	font-style: normal;
	}
.test, .ref { border: 1px solid orange;  margin: 20px;  padding: 10px; width: 390px; font: 24px/1 csstest_asciiregular; white-space:nowrap; }
/* the CSS above is not part of the test */
</style>
</head>
<body>
<div id='log'></div>
<div class="test" lang="de"><span id="testspan">Eingabe&#x00AD;verarbeitungs&#x00AD;funktionen</span></div>
<div class="ref" lang="de"><span id="refspan">Eingabeverarbeitungsfunktionen</div>
<script>
test(function() {
assert_equals(document.getElementById('testspan').offsetWidth, document.getElementById('refspan').offsetWidth);
}, "words should be the same length");
</script>
</body>
</html>
